code/requirements-v2.txt
absl-py==1.4.0
alabaster==0.7.12
anaconda-client==1.7.2
anaconda-navigator==1.7.0
anaconda-project==0.8.3
appdirs==1.4.3
asn1crypto==0.24.0
astroid==2.2.5
astropy==3.2.1
atomicwrites==1.3.0
attrs==19.1.0
Automat==0.7.0
Babel==2.7.0
backcall==0.1.0
backports.os==0.1.1
backports.shutil-get-terminal-size==1.0.0
beautifulsoup4==4.7.1
bitarray==0.9.3
bkcharts==0.2
blaze==0.11.3
bleach==3.1.0
bokeh==1.2.0
boto==2.49.0
Bottleneck==1.2.1
cachetools==4.2.4
certifi==2019.6.16
cffi==1.12.3
chardet==3.0.4
Click==7.0
cloudpickle==1.2.1
clyent==1.2.2
colorama==0.4.1
conda==4.7.12
conda-build==3.17.6
conda-package-handling==1.6.0
conda-verify==3.1.1
constantly==15.1.0
contextlib2==0.5.5
cryptography==2.7
cycler==0.10.0
Cython==0.29.12
cytoolz==0.10.0
dask==2.1.0
dataclasses==0.8
datashape==0.5.4
decorator==4.4.0
defusedxml==0.6.0
distributed==2.1.0
docutils==0.14
entrypoints==0.3
et-xmlfile==1.0.1
fastcache==1.1.0
filelock==3.0.12
Flask==1.1.1
Flask-Cors==3.0.7
future==0.17.1
gevent==1.4.0
glob2==0.7
gmpy2==2.0.8
google-auth==2.22.0
google-auth-oauthlib==0.4.6
googledrivedownloader==0.4
greenlet==0.4.15
grpcio==1.48.2
h5py==2.9.0
heapdict==1.0.0
html5lib==1.0.1
hyperlink==18.0.0
idna==2.8
imageio==2.5.0
imagesize==1.1.0
importlib-metadata==4.8.3
incremental==17.5.0
ipykernel==5.1.1
ipython==7.6.1
ipython_genutils==0.2.0
ipywidgets==7.5.0
isodate==0.6.0
isort==4.3.21
itsdangerous==1.1.0
jdcal==1.4.1
jedi==0.13.3
jeepney==0.4
Jinja2==2.10.1
joblib==0.13.2
json5==0.8.4
jsonschema==3.0.1
jupyter==1.0.0
jupyter-client==5.3.1
jupyter-console==6.0.0
jupyter-core==4.5.0
jupyterlab==1.0.2
jupyterlab-launcher==0.13.1
jupyterlab-server==1.0.0
keyring==18.0.0
kiwisolver==1.1.0
lazy-object-proxy==1.4.1
libarchive-c==2.8
lief==0.9.0
llvmlite==0.29.0
locket==0.2.0
lxml==4.3.4
Markdown==3.3.7
MarkupSafe==1.1.1
matplotlib==3.1.0
mccabe==0.6.1
mistune==0.8.4
mkl-fft==1.0.12
mkl-random==1.0.2
mkl-service==2.0.2
mock==3.0.5
more-itertools==7.0.0
mpmath==1.1.0
msgpack==0.6.1
multipledispatch==0.6.0
navigator-updater==0.1.0
nbconvert==5.5.0
nbformat==4.4.0
networkx==2.3
nltk==3.4.4
nose==1.3.7
notebook==6.0.0
numba==0.45.0
numexpr==2.6.9
numpy==1.16.4
numpydoc==0.9.1
oauthlib==3.2.2
odo==0.5.1
olefile==0.46
openpyxl==2.6.2
packaging==19.0
pandas==0.24.2
pandocfilters==1.4.2
parso==0.5.0
partd==1.0.0
path.py==12.0.1
pathlib2==2.3.4
patsy==0.5.1
pep8==1.7.1
pexpect==4.7.0
pickleshare==0.7.5
Pillow==6.1.0
pkginfo==1.5.0.1
pluggy==0.12.0
ply==3.11
prometheus-client==0.7.1
prompt-toolkit==2.0.9
protobuf==3.19.6
psutil==5.6.3
ptyprocess==0.6.0
py==1.8.0
pyasn1==0.4.4
pyasn1-modules==0.2.2
pycodestyle==2.5.0
pycolmap==0.3.0
pycosat==0.6.3
pycparser==2.19
pycrypto==2.6.1
pycurl==7.43.0.3
pyflakes==2.1.1
Pygments==2.4.2
pylint==2.3.1
pyodbc==4.0.26
pyOpenSSL==19.0.0
pyparsing==2.4.0
pyrsistent==0.14.11
PySocks==1.7.0
pytest==5.0.1
pytest-arraydiff==0.3
pytest-astropy==0.5.0
pytest-doctestplus==0.3.0
pytest-openfiles==0.3.2
pytest-remotedata==0.3.1
python-dateutil==2.8.0
pytz==2019.1
PyWavelets==1.0.3
PyYAML==5.1.1
pyzmq==18.0.0
QtAwesome==0.5.7
qtconsole==4.5.1
QtPy==1.8.0
rdflib==5.0.0
requests==2.22.0
requests-oauthlib==2.0.0
rope==0.14.0
rsa==4.9
ruamel_yaml==0.15.46
scikit-image==0.15.0
scikit-learn==0.21.2
scipy==1.3.0
seaborn==0.9.0
SecretStorage==3.1.1
Send2Trash==1.5.0
service-identity==17.0.0
simplegeneric==0.8.1
singledispatch==3.4.0.3
six==1.12.0
snowballstemmer==1.9.0
sortedcollections==1.1.2
sortedcontainers==2.1.0
soupsieve==1.8
Sphinx==2.1.2
sphinxcontrib-applehelp==1.0.1
sphinxcontrib-devhelp==1.0.1
sphinxcontrib-htmlhelp==1.0.2
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.2
sphinxcontrib-serializinghtml==1.1.3
sphinxcontrib-websupport==1.1.2
spyder==3.3.6
spyder-kernels==0.5.1
SQLAlchemy==1.3.5
statsmodels==0.10.0
sympy==1.4
tables==3.5.2
tblib==1.4.0
tensorboard==2.10.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
terminado==0.8.2
testpath==0.4.2
toolz==0.10.0
torch-scatter==2.0.7
tornado==6.0.3
tqdm==4.32.1
traitlets==4.3.2
Twisted==18.7.0
typed-ast==1.3.4
typing==3.6.2
typing_extensions==4.1.1
unicodecsv==0.14.1
urllib3==1.24.2
wcwidth==0.1.7
webencodings==0.5.1
Werkzeug==2.0.3
widgetsnbextension==3.5.0
wrapt==1.11.2
wurlitzer==1.0.2
xlrd==1.2.0
XlsxWriter==1.1.8
xlwt==1.3.0
yacs==0.1.8
zict==1.0.0
zipp==0.5.1
zope.interface==4.5.0
